pyproject.toml
[project]
name = "media_core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.2",
    "djangorestframework>=3.16.0",
    "flower>=2.0.1",
    "minio>=7.2.15",
    "pika>=1.3.2",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "redis>=6.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-django>=4.11.1",
    "pytest-dotenv>=0.5.2",
]
